Extract view reducer
