+++ test/selectors/report-tree_spec.js
@@ -1,6 +1,6 @@
 import { fromJS } from "immutable";
 import getSequenceTree, { getAnswerTrees, getQuestionTrees, getPageTrees, getSectionTrees, getActivityTrees } from "../../js/selectors/report-tree";
-import { FULL_REPORT, DASHBOARD } from "../../js/reducers";
+import { FULL_REPORT, DASHBOARD } from "../../js/reducers/view-reducer";
 
 describe("report tree selectors", () => {
   const state = ({ questionVisible = true, hideSectionNames = true }) => fromJS({
